refactor(herdr): split event parsing and streaming
