+++ crates/board-herdr/src/events/stream.rs
@@ -1,36 +1,17 @@
-//! Event streaming over a dedicated herdr connection.
-//!
-//! `events.subscribe` must run on its own persistent socket (never the
-//! request/response connection). This module opens that socket, subscribes,
-//! and exposes a blocking [`Iterator`] of [`HerdrEvent`].
-//!
-//! ## Subscription quirk (verified live, protocol 17)
-//! A `pane.agent_status_changed` subscription **requires a concrete `pane_id`**
-//! — herdr validates the pane exists and rejects a wildcard/missing id with
-//! `internal_error`. So the daemon must build one subscription per pane it
-//! wants status for (see [`watch_subscriptions`]) and re-subscribe (or
-//! reconnect) as it starts new agents. `pane.exited` / `pane.closed` are global
-//! and take no `pane_id`.
-//!
-//! Emitted event lines use the `EventEnvelope` shape
-//! `{"event":"<kind>","data":{"type":"<kind>",...}}` with **underscore** kind
-//! names (`pane_agent_status_changed`, `pane_exited`), whereas *subscription*
-//! entries use **dotted** names (`pane.agent_status_changed`). Both are handled
-//! here.
-
 use std::collections::VecDeque;
 use std::io::{BufRead, BufReader, Write};
 use std::os::unix::net::UnixStream;
 use std::path::{Path, PathBuf};
 use std::time::{Duration, Instant};
 
-use serde::Deserialize;
 use serde_json::{json, Value};
 
 use crate::envelope::{Request, Response};
 use crate::error::{HerdrError, Result};
 use crate::transport::{self, connect_with_deadline, SocketDeadlines};
-use crate::types::AgentStatus;
+
+use super::backoff::Backoff;
+use super::parse::{parse_event_line, HerdrEvent};
 
 /// One subscription entry for `events.subscribe`.
 #[derive(Debug, Clone, PartialEq, Eq)]
@@ -69,136 +50,6 @@ pub fn watch_subscriptions(pane_ids: &[String]) -> Vec<Subscription> {
     let mut subs = vec![Subscription::pane_exited(), Subscription::pane_closed()];
     subs.extend(pane_ids.iter().map(|p| Subscription::agent_status(p)));
     subs
-}
-
-/// A decoded event. Tolerant: unknown event kinds become [`HerdrEvent::Other`]
-/// carrying the raw line, and unknown fields are ignored.
-#[derive(Debug, Clone, PartialEq)]
-pub enum HerdrEvent {
-    /// `pane_agent_status_changed`.
-    AgentStatusChanged {
-        pane_id: String,
-        workspace_id: Option<String>,
-        status: AgentStatus,
-        agent: Option<String>,
-    },
-    /// `pane_exited` or `pane_closed` — the pane is gone.
-    PaneExited {
-        pane_id: String,
-        workspace_id: Option<String>,
-    },
-    /// Any other event line (raw envelope preserved).
-    Other(Value),
-}
-
-#[derive(Deserialize)]
-struct StatusFields {
-    #[serde(default)]
-    pane_id: Option<String>,
-    #[serde(default)]
-    workspace_id: Option<String>,
-    #[serde(default)]
-    agent_status: Option<AgentStatus>,
-    #[serde(default)]
-    agent: Option<String>,
-}
-
-#[derive(Deserialize)]
-struct PaneFields {
-    #[serde(default)]
-    pane_id: Option<String>,
-    #[serde(default)]
-    workspace_id: Option<String>,
-}
-
-/// Parse one raw NDJSON line into an event, or `None` if it is not an event
-/// (e.g. the `subscription_started` ack, an error/ack line, or blank).
-pub fn parse_event_line(line: &str) -> Option<HerdrEvent> {
-    let line = line.trim();
-    if line.is_empty() {
-        return None;
-    }
-    let value: Value = serde_json::from_str(line).ok()?;
-    // Skip request/response envelopes (acks, errors) — they carry an `id`.
-    if value.get("result").is_some() || value.get("error").is_some() {
-        return None;
-    }
-    // Event body: prefer the nested `data` object, else the value itself.
-    let data = match value.get("data") {
-        Some(d) if d.is_object() => d,
-        _ => &value,
-    };
-    // The kind lives in `data.type` (underscore names) on some herdr builds and
-    // in the top-level `event` key (dotted names) on others (verified live on
-    // protocol 17: {"event":"pane.agent_status_changed","data":{...}} with
-    // no data.type). Accept both, normalized to underscores.
-    let kind = data
-        .get("type")
-        .and_then(Value::as_str)
-        .or_else(|| value.get("event").and_then(Value::as_str))?
-        .replace('.', "_");
-    match kind.as_str() {
-        "pane_agent_status_changed" => {
-            let f: StatusFields = serde_json::from_value(data.clone()).ok()?;
-            let pane_id = f.pane_id?;
-            Some(HerdrEvent::AgentStatusChanged {
-                pane_id,
-                workspace_id: f.workspace_id,
-                status: f.agent_status.unwrap_or(AgentStatus::Unknown),
-                agent: f.agent,
-            })
-        }
-        "pane_exited" | "pane_closed" => {
-            let f: PaneFields = serde_json::from_value(data.clone()).ok()?;
-            let pane_id = f.pane_id?;
-            Some(HerdrEvent::PaneExited {
-                pane_id,
-                workspace_id: f.workspace_id,
-            })
-        }
-        _ => Some(HerdrEvent::Other(value)),
-    }
-}
-
-/// Retry policy for [`HerdrEvents::connect_with_retry`].
-#[derive(Debug, Clone)]
-pub struct Backoff {
-    pub initial: Duration,
-    pub max: Duration,
-    pub multiplier: f64,
-    /// `None` = retry forever (daemon default); `Some(n)` = give up after `n`
-    /// failed attempts and return the last error.
-    pub max_retries: Option<usize>,
-}
-
-impl Default for Backoff {
-    fn default() -> Backoff {
-        Backoff {
-            initial: Duration::from_millis(200),
-            max: Duration::from_secs(5),
-            multiplier: 2.0,
-            max_retries: None,
-        }
-    }
-}
-
-impl Backoff {
-    /// A bounded policy (useful for tests).
-    pub fn bounded(max_retries: usize) -> Backoff {
-        Backoff {
-            max_retries: Some(max_retries),
-            ..Backoff::default()
-        }
-    }
-
-    fn next_delay(&self, current: Duration) -> Duration {
-        let next = current.mul_f64(self.multiplier);
-        if next > self.max {
-            self.max
-        } else {
-            next
-        }
-    }
 }
 
 /// A persistent event-stream connection. Iterating yields decoded events until
